+++ db/migrate/20251016004734_extend_content_plain_response.rb
@@ -1,0 +1,5 @@
+class ExtendContentPlainResponse < ActiveRecord::Migration[8.0]
+  def change
+    change_column :plain_responses, :content, :text
+  end
+end